refactor(ui): move type to types directory
