--- ui/src/timeMachine/constants/index.ts
@@ -1,2 +1,0 @@
-// TODO: Move me
-export type TimeMachineID = 'de' | 'veo' | 'alerting'
